fix: bug in api call to update a resource in dsc
